requirements.txt
# Flask 웹 프레임워크
flask==2.3.3
flask-cors==6.0.0
werkzeug==3.1.3

# 환경변수 관리
python-dotenv==1.1.0

# HTTP 요청 라이브러리
requests==2.32.3

# AWS SDK
boto3==1.38.22
botocore==1.38.22

# Google Cloud Speech API
google-cloud-speech==2.32.0
google-api-core==2.24.2
google-auth==2.40.2
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
proto-plus==1.26.1
protobuf==6.31.0

# 컴퓨터 비전 및 이미지 처리
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
numpy==1.26.3
pillow==10.2.0

# 딥러닝 프레임워크 (PyTorch)
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

# 얼굴 인식
facenet-pytorch==2.6.0

# 감정 분석 라이브러리
emotiefflib==1.0

# 유틸리티 라이브러리들
tqdm==4.67.1
pyyaml==6.0.2
coloredlogs==15.0.1
humanfriendly==10.0

# 기타 의존성
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
urllib3==1.26.20
six==1.17.0
python-dateutil==2.9.0.post0
jmespath==1.0.1
s3transfer==0.13.0
packaging==25.0
filelock==3.13.1
fsspec==2024.6.1
huggingface-hub==0.31.2
safetensors==0.5.3
sympy==1.13.3
networkx==3.2.1
mpmath==1.3.0
typing-extensions==4.12.2